github+Docker: bump and fix go version to 1.16.9
There's a race condition in the h2_bundle.go of go 1.16.10 and later. The
issue https://github.com/golang/go/issues/51799 mentions that this might
be fixed in go 1.19, so we'll need to wait for that.
We make sure we build our docker images with go 1.16.9 to not run into
the issue in the wild.