pricesrpc: add http_request_text to GetPrice
In this commit, we add the http_request_text to the `GetPrice` RPC call.
This'll contain the fully serialized HTTP request, which'll permit the
back end to check things like the set of headers or request body, which
may be useful in more fine grained pricing functions.

One example is: when proxying requests to an LLM platform provider, the
backed may want to charge more based on the length of the
context/message and/or which model is being used.