proxy: set Content-Type for gRPC requests
Prior we would not set the Content-Type header for
grpc request, which on failure events would cause
the response to be invalid gRPC responses, as the
grpc code path in sendDirectResponse expects the
Content-Type header to be set.

This change fixes the problem by setting the
Content-Type header to application/grpc for gRPC
requests.